Update testFDC3Service to exampleFDC3Service.
- Update the readme
- Update the service to simplify
- Add the service dependancies
- Rename the service
