+++ src/services/exampleFDC3/exampleFDC3.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>ExampleFDC3</title>
+</head>
+
+<body>
+    <h1>ExampleFDC3</h1>
+</body>
+<script src="exampleFDC3Service.js"></script>
+
+</html>
