--- src/services/testFDC3/testFDC3.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>TestFDC3</title>
-</head>
-
-<body>
-    <h1>TestFDC3</h1>
-</body>
-<script src="testFDC3Service.js"></script>
-
-</html>
